+++ iter.h
@@ -13,9 +13,9 @@
 struct block {
 	struct list_head node;
 	unsigned char *data;
-	unsigned int size;
-	unsigned int alloc;
-	unsigned int nl;
+	long size;
+	long alloc;
+	long nl;
 };
 
 static inline struct block *BLOCK(struct list_head *item)
@@ -26,25 +26,25 @@ static inline struct block *BLOCK(struct list_head *item)
 struct block_iter {
 	struct block *blk;
 	struct list_head *head;
-	unsigned int offset;
+	long offset;
 };
 
 struct lineref {
 	const unsigned char *line;
-	unsigned int size;
+	long size;
 };
 
 void block_iter_normalize(struct block_iter *bi);
-unsigned int block_iter_eat_line(struct block_iter *bi);
-unsigned int block_iter_next_line(struct block_iter *bi);
-unsigned int block_iter_prev_line(struct block_iter *bi);
-unsigned int block_iter_bol(struct block_iter *bi);
-unsigned int block_iter_eol(struct block_iter *bi);
-void block_iter_back_bytes(struct block_iter *bi, unsigned int count);
-void block_iter_skip_bytes(struct block_iter *bi, unsigned int count);
-void block_iter_goto_offset(struct block_iter *bi, unsigned int offset);
-void block_iter_goto_line(struct block_iter *bi, unsigned int line);
-unsigned int block_iter_get_offset(const struct block_iter *bi);
+long block_iter_eat_line(struct block_iter *bi);
+long block_iter_next_line(struct block_iter *bi);
+long block_iter_prev_line(struct block_iter *bi);
+long block_iter_bol(struct block_iter *bi);
+long block_iter_eol(struct block_iter *bi);
+void block_iter_back_bytes(struct block_iter *bi, long count);
+void block_iter_skip_bytes(struct block_iter *bi, long count);
+void block_iter_goto_offset(struct block_iter *bi, long offset);
+void block_iter_goto_line(struct block_iter *bi, long line);
+long block_iter_get_offset(const struct block_iter *bi);
 
 bool block_iter_is_bol(const struct block_iter *bi);
 
@@ -55,6 +55,6 @@ static inline bool block_iter_is_eof(struct block_iter *bi)
 
 void fill_line_ref(struct block_iter *bi, struct lineref *lr);
 void fill_line_nl_ref(struct block_iter *bi, struct lineref *lr);
-unsigned int fetch_this_line(const struct block_iter *bi, struct lineref *lr);
+long fetch_this_line(const struct block_iter *bi, struct lineref *lr);
 
 #endif
