+++ obuf.h
@@ -6,7 +6,7 @@
 
 struct output_buffer {
 	char buf[8192];
-	unsigned int count;
+	long count;
 
 	// number of characters scrolled (x direction)
 	unsigned int scroll_x;
