+++ selection.h
@@ -5,13 +5,13 @@
 
 struct selection_info {
 	struct block_iter si;
-	unsigned int so;
-	unsigned int eo;
+	long so;
+	long eo;
 	bool swapped;
 };
 
 void init_selection(struct selection_info *info);
-unsigned int prepare_selection(void);
+long prepare_selection(void);
 int get_nr_selected_lines(struct selection_info *info);
 int get_nr_selected_chars(struct selection_info *info);
